refactor(monolith-public): move notes chat to /app/notes, lazy graph toggle, drop CSP
Move the public notes chat under /app/notes to sit alongside the other apps
(ships/stars/hikes/dr-jobs). Chat is the default view with an in-page Chat |
Graph toggle; the graph view (heavy KnowledgeGraph + d3) and its data are
lazy-loaded via dynamic import + a same-origin /app/notes/graph proxy only when
the visitor switches to it. node_touched highlighting (touchedIds) is preserved.

Old routes 308-redirect to /app/notes (/notes and /notes/graph). The note-body
proxy moves to /app/notes/body. Notes leaves the top-level nav and joins the
APPS dropdown. Homepage + 404 CTAs, llms.txt, and sitemap.xml point at the new
URL.

Remove the Phase 4c CSP entirely (deferred to a later hardening pass): drop
kit.csp from svelte.config.js and delete src/lib/csp.js + csp.test.js. The
markdown sanitization/escaping and its XSS tests stay (the real protection). No
CSP header or meta is set by the app.

Bumps the monolith-public chart to 0.6.8 (Chart.yaml + application.yaml
targetRevision) since the frontend image changes.
