+++ projects/monolith/frontend/src/routes/+layout.svelte
@@ -24,16 +24,16 @@
   // public.jomcgi.dev/* → /public/* and private.jomcgi.dev/* → /private/*
   // internally, but $page.url reflects the *browser* URL. So:
   // - /review (private host) → "review"
-  // - /notes (any host) → "notes"
   // - /engineering (any host) → "engineering"
   // - /cv (any host) → "cv"
   // - any other URL on public.jomcgi.dev → "home"
   // - everything else → no active state
+  // Notes now lives under /app/notes, which suppresses the layout nav (it
+  // renders its own with route="notes"), so it needs no detection here.
   let activeRoute = $derived.by(() => {
     const host = $page.url.hostname;
     const path = $page.url.pathname;
     if (path === "/review" || path.startsWith("/review/")) return "review";
-    if (path === "/notes" || path.startsWith("/notes/")) return "notes";
     if (path === "/engineering" || path.startsWith("/engineering/")) {
       return "engineering";
     }
