+++ projects/monolith/frontend/src/routes/public/llms.txt/+server.js
@@ -17,8 +17,7 @@ const BODY = `# Joe McGinley
 ## Pages
 
 - [CV](${PUBLIC_BASE}/cv): full work history and project case studies
-- [Notes](${PUBLIC_BASE}/notes): chat with my public knowledge graph
-- [Graph](${PUBLIC_BASE}/notes/graph): the public knowledge graph, browsable
+- [Notes](${PUBLIC_BASE}/app/notes): chat with my public knowledge graph, or switch to the graph view to browse it
 - [Home](${PUBLIC_BASE}/): overview and homelab status
 
 ## Elsewhere
