+++ projects/monolith/frontend/src/routes/public/notes/+page.server.js
@@ -1,41 +1,8 @@
-import { error } from "@sveltejs/kit";
-import {
-  NOTES_PAGE_CACHE_CONTROL,
-  versionedEtag,
-} from "../../../lib/cache-headers.js";
+import { redirect } from "@sveltejs/kit";
 
-// The localhost fallback is the established convention across every public
-// +page.server.js proxy (ships/stars/notes/body); prod sets API_BASE via
-// values.yaml.
-// nosemgrep: sveltekit-server-hardcoded-api-base-fallback
-const API_BASE = process.env.API_BASE || "http://localhost:8000";
-
-// The public Turnstile site key gates the chat landing (ADR 005). It is PUBLIC
-// by design (it identifies the widget, not a credential); the Turnstile *secret*
-// never enters the frontend, only the FastAPI backend. Read from the
-// environment, mirroring how the chat route exposes it.
-const TURNSTILE_SITE_KEY = process.env.TURNSTILE_SITE_KEY || "";
-
-// Mirrors private/notes/+page.server.js, but hits the visibility-filtered
-// public endpoint. The backend enforces `visibility = 'public'` on both
-// nodes and edges (both-ends-public), so we just pass the payload through.
-// The public notes landing is now the neo-brutalist chat front door (ADR 005):
-// the graph ships with the page for the deep-dive overlay, and the site key
-// gates the "start chatting" challenge.
-export async function load({ fetch, setHeaders }) {
-  const res = await fetch(`${API_BASE}/api/knowledge/public/graph`, {
-    signal: AbortSignal.timeout(10_000),
-  });
-  if (!res.ok) {
-    throw error(503, "graph unavailable");
-  }
-
-  const headers = { "cache-control": NOTES_PAGE_CACHE_CONTROL };
-  const etag = versionedEtag(res.headers?.get?.("etag"));
-  if (etag) headers.etag = etag;
-  const lastModified = res.headers?.get?.("last-modified");
-  if (lastModified) headers["last-modified"] = lastModified;
-  setHeaders(headers);
-
-  return { graph: await res.json(), turnstileSiteKey: TURNSTILE_SITE_KEY };
+// The notes chat + graph moved to /app/notes to sit alongside the other apps
+// (ships/stars/hikes/dr-jobs). Permanent-redirect the old front door so existing
+// links, the homepage CTA, and any bookmarks do not 404.
+export function load() {
+  throw redirect(308, "/app/notes");
 }
