+++ projects/monolith/frontend/svelte.config.js
@@ -1,6 +1,5 @@
 import adapter from "@sveltejs/adapter-node";
 import { mdsvex } from "mdsvex";
-import { cspDirectives } from "./src/lib/csp.js";
 
 const config = {
   extensions: [".svelte", ".svx"],
@@ -12,17 +11,11 @@ const config = {
     adapter: adapter({
       out: process.env.SVELTE_OUT_DIR || "build",
     }),
-    // Content-Security-Policy (ADR 005 layer 8, Phase 4c). mode "auto" nonces
-    // SvelteKit's own inline bootstrap scripts under SSR, so script-src can stay
-    // strict (no 'unsafe-inline') and untrusted public-chat output cannot run
-    // script. The host allow-list and the style-relaxation rationale live in
-    // src/lib/csp.js. This is global config (one svelte.config.js drives both
-    // the public and private builds), so the directives are audited against all
-    // routes, not just the public chat.
-    csp: {
-      mode: "auto",
-      directives: cspDirectives,
-    },
+    // No Content-Security-Policy is set by the app. The markdown renderer
+    // (components/notes/markdown.js) HTML-escapes untrusted public-chat output
+    // and note bodies and emits no raw HTML, which is the real XSS protection
+    // (covered by markdown.test.js). A CSP hardening layer is deferred to a
+    // later pass.
   },
 };
